requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
tortoise-orm==0.25.1
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
